% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_fat.R
\docType{data}
\name{blood_fat}
\alias{blood_fat}
\title{Blood fat content}
\format{A data frame with 400 rows and 3 variables:
\describe{
  \item{id}{An index}
  \item{weight}{weight in kilograms}
  \item{age}{age in years}
  \item{fat}{Blood fat content}
}}
\usage{
blood_fat
}
\description{
A dataset from http://people.sc.fsu.edu/~jburkardt/datasets/regression/x09.txt
}
\keyword{datasets}

