% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garden.R
\docType{data}
\name{garden}
\alias{garden}
\title{ozone concentration in 3 different gardens.}
\format{A tibble with 10 rows and 3 variables:
\describe{
  \item{gardenA}{ozone concentration for gardenA}
  \item{gardenB}{ozone concentration for gardenB}
  \item{gardenC}{ozone concentration for gardenC}
}}
\usage{
garden
}
\description{
A (fake) dataset from Michael Crawley
}
\keyword{datasets}

