% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mousew.R
\docType{data}
\name{mousew}
\alias{mousew}
\title{Weights of mice.}
\format{A data frame with 400 rows and 3 variables:
\describe{
  \item{sex}{gender, Male (M) or Female (F)}
  \item{strain}{strain name, either C57Bl/6J or DBA/2J}
  \item{weight}{weight, in grams}
}}
\usage{
mousew
}
\description{
A (fake) dataset containing the weights of 12 week old male
and female mice from two different strains (C57Bl/6J and DBA/2J)
}
\keyword{datasets}

