% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mousew_summary.R
\docType{data}
\name{mousew_summary}
\alias{mousew_summary}
\title{Weights of mice.}
\format{A data frame with 4 rows and 5 variables:
\describe{
  \item{sex}{gender, Male (M) or Female (F)}
  \item{strain}{strain name, either C57Bl/6J or DBA/2J}
  \item{median_weight}{median mouse weight}
  \item{mean_weight}{mean mouse weight}
  \item{sd_weight}{standard deviation of mouse weight}
}}
\usage{
mousew_summary
}
\description{
A (fake) dataset containing the weights of 12 week old male
and female mice from two different strains (C57Bl/6J and DBA/2J)
}
\keyword{datasets}

