% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wings.R
\docType{data}
\name{wings}
\alias{wings}
\title{Size of insect wings on 2 different locations.}
\format{A tibble with 120 rows and 2 variables:
\describe{
  \item{size}{size, in mm}
  \item{location}{A or B}
}}
\usage{
wings
}
\description{
A (fake) dataset from Michael Crawley
}
\keyword{datasets}

